unified/requirements.txt
# Python 3.11 dependencies - fixed versions for stability
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
pyreadstat==1.2.7
pandas==2.1.4
numpy==1.26.2
mammoth==1.6.0
python-docx==1.1.0
gunicorn==21.2.0
